Feature/loginggoogle

+++ src/main/java/com/helpinghandslocation/helpinghandslocation/repositories/TypeRepository.java
@@ -3,5 +3,5 @@ package com.helpinghandslocation.helpinghandslocation.repositories;
 import com.helpinghandslocation.helpinghandslocation.models.Type;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeRepository extends JpaRepository<Type, Integer> {
+public interface TypeRepository extends JpaRepository<Type, Long> {
 }
